Move test suite into request_handler module
